(add) useMovieSearchData hook
