+++ js/main/hook/useMovieData.js
@@ -23,19 +23,10 @@ const useMovieData = () => {
     return res.results;
   };
 
-  const getSearchedMoviesData = async (page = 1, query) => {
-    const queryToUnicode = encodeURIComponent(query);
-    const res = await get(
-      `${TMDB_API_URL}/search/movie?query=${queryToUnicode}&include_adult=false&language=ko-KR&page=${page}&api_key=${API_KEY}`
-    );
-    return res.results;
-  };
-
   return {
     getPopularMovies,
     getTopRatedMovies,
     getUpcomingMovies,
-    getSearchedMoviesData,
   };
 };
